Rename to jsx, remove old code
